Run the Chrome video-contrast tests with the Chrome suites
The two tests asserting Chrome's no-video behavior lived in the
focused Firefox suite, which now runs in a Firefox-only CI job where
Chrome is not installed. Move them to their own file listed under
test-js-async so the Chrome job runs them, and drop the Chrome
install from the Firefox job again.
