+++ tests/js/async/firefox.test.js
@@ -152,46 +152,6 @@ describe('JS Firefox', () => {
     }
   });
 
-  test('required video on chrome fails with an actionable error', async () => {
-    const bro = await browser.start({ headless: true });
-    try {
-      const vibe = await bro.page();
-      await vibe.go(baseURL);
-      await assert.rejects(
-        () => vibe.context.recording.start({ video: true, path: null }),
-        (err) => /not supported by this browser/.test(err.message),
-        'Chrome should explain video is unsupported and name the firefox install'
-      );
-    } finally {
-      await bro.stop();
-    }
-  });
-
-  test('omitted video on chrome records a trace without a video track', async () => {
-    const bro = await browser.start({ headless: true });
-    try {
-      const vibe = await bro.page();
-      await vibe.go(baseURL);
-
-      await vibe.context.recording.start({ path: null });
-      await vibe.find('a[href="/login"]', { timeout: 5000 });
-      const result = await vibe.context.recording.stop();
-      assert.ok(result.bytes.length > 0, 'Recording should still deliver');
-      assert.ok(result.videoUnavailable, 'Result should say why there is no video');
-
-      const { extractedDir, cleanup } = unzipRecording(result.bytes);
-      try {
-        const files = fs.readdirSync(extractedDir);
-        assert.ok(files.some((f) => f.endsWith('.trace')), 'Trace should be present');
-        assert.ok(!files.includes('video'), 'No video entries on an engine without support');
-      } finally {
-        cleanup();
-      }
-    } finally {
-      await bro.stop();
-    }
-  });
-
   test('unknown browser is rejected with a clear error', (t) => {
     const bin = process.env.VIBIUM_BIN_PATH;
     if (!bin) return skipOrFail(t, 'VIBIUM_BIN_PATH not set');
